==> 03_revisions/05_mg-analysis.ipynb <==
import os
from pathlib import Path
import sys
import scanpy as sc
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np
import numpy as np
import pandas as pd
from scipy.stats import mannwhitneyu
from itertools import product
import session_info

import matplotlib as mpl
mpl.rcParams['pdf.fonttype'] = 42

plt.rcParams['figure.dpi'] = 500

input_dir = '/path/to/data/h5ad/'
input_adata = os.path.join(input_dir, '02_mg-annotated.h5ad')
output_dir = input_dir

adata = sc.read_h5ad(input_adata)

# Salvador et al., gene list, DEGs

homeostatic_microglia_genes = [
    'P2ry12', 'Siglech', 'Selplg', 'Cst3', 'Tmem119', 'Jun', 'Socs3', 'Adgrg1', 'Ccr5',
    'Jund', 'Col27a1', 'Vsir', 'Klf4', 'Crybb1', 'Nrip1', 'Cebpd', 'Txnip', 'Rhoh', 'Ssh2',
    'Btg2', 'Ivns1abp', 'Serinc3', 'Rhob', 'Arhgap5', 'Zfhx3', 'Il6ra', 'Egr1', 'Ier5',
    'Ifngr1', 'E230029C05Rik', 'Frmd4a', 'Pag1', 'Srgap2', 'Tanc2', 'Slc2a5', 'P2ry13',
    'Klf2', 'Zfp36', 'Dock10', 'Slco2b1', 'Ptgs1', 'Csf1r', 'Sox4', 'Tmx4', 'Junb', 'Srsf5',
    'Hpgd', 'Sall1', 'Nav3', 'Glul', 'Tmem173', 'Tgfbr1', 'Cd74', 'Tsc22d3', 'Dock4', 'Fgd2',
    'Pik3r1', 'Jmjd1c', 'Git2', 'St3gal6', 'Elmo1', 'Susd3', 'Entpd1', 'Cd164', 'Pde3b',
    'Cx3cr1', 'Adrb2', 'Csmd3', 'Zfp36l2', 'Cmtm6', 'Kctd12', 'H2-Ab1', 'Abca9', 'Gm34455',
    'Lpcat2', 'Mef2a', 'Mertk', 'Ighm', 'Ddx5', 'Lrba', 'Ankrd44', 'Tra2a', 'Picalm',
    'Slc16a6', 'Atp8a2', 'Adap2', 'Xist', 'Sbno2', 'Adrb1', 'Fchsd2', 'Lifr', 'Lair1',
    'Tmcc3', '4632427E13Rik', 'Fscn1', 'Nav2', 'Tm6sf1', 'Numb', 'G530011O06Rik', 'Malat1',
    '2810013P06Rik', 'Slc1a3', 'Gm47664', 'Sparc', 'Slc29a3', 'Il10ra', 'Wnk1', 'Celf2',
    'Sult1a1', 'March1', 'Maf', 'Rapgef5', 'Ubc', 'Dennd4a', 'Serpinf1', 'Fam102b',
    'Hnrnph1', 'Zbtb20', 'Scamp2', 'Klf3', 'Pou2f2', 'Otulinl', 'Ltc4s', 'Dusp6', 'Clec4a2',
    'Selenop', 'Marcks', 'F11r', 'Cfh', 'Fos', 'Tuba1a', 'Arid1a', 'Cttnbp2nl', 'Lag3',
    'Dnajb1', 'Gpr34', 'Inpp5d', 'Dapp1', 'Cysltr1', 'Rcsd1', 'Hpgds', 'Pld4', 'Luc7l2',
    '1700017B05Rik', 'Crlf3', 'Wsb1', 'Clk1', 'Ikzf1', 'Adora3', 'Tpst2', 'Frmd4b', 'Rnase4',
    'Zfp652', 'Tsix', 'Mycbp2', 'Foxn3', 'Unc93b1', 'Il7r', 'Chd9', 'Rock2', 'Btg1', 'Mbnl1',
    'Pnisr', 'Fosb', 'Dip2b', 'Gm15503', 'Rogdi', 'Epb41l2', 'Pld1', 'Rassf2', 'Mir99ahg',
    'Runx1', 'Ophn1', 'Il16', 'Kcnk6', 'Rbm5', 'Pwwp2a', 'Gpr155', 'Phf14', 'Mgat4a', 'Acox3',
    'Snrnp70', 'Ppp1r9a', 'Gm2629', 'Irf2bpl', 'Rbm39', 'Snx18', 'Akap8l', 'Ccnl1', 'Ctc1',
    'Hexb', 'St3gal5', 'Usp2', 'Bmp2k', 'Eif4a2', 'Eef2k', 'Etv3', 'Zbtb18', 'Sipa1',
    'Ints6l', 'Nfat5', 'Tent5a', 'Mtdh', 'Smad7', 'Zcchc7', 'Zfp710', 'Klf7', 'Hps3', 'Cd33',
    'Arglu1', 'Rtn1', 'Soga1', 'Kdm6b', 'Pan3', 'P2ry6', 'Hps4', 'Adap2os', '0610040J01Rik',
    'Rgs2', 'Sall3', 'Ogfrl1', 'Lpar6', 'Pip4k2a', 'Filip1l', 'Tacc1', 'Qk', 'Akirin2',
    'Ecscr', 'Gal3st4', 'Golm1', 'Slc12a9', 'Smap2', 'Wdr44', 'Gm43813', 'Fkbp5', 'Cmtm7',
    'Tnrc6b', 'Zfp36l1', 'Maml3', 'Tab2', 'Tjp1', 'Nfam1', 'Gadd45g', 'Skil', 'Ppcdc', 'Dst',
    'Pla2g15', 'Pnp', 'Rps6ka1', 'Capn3', 'Ier2', 'Ubn2', 'Camk2n1', 'Kmt2e', 'Gm29488',
    'Map4k4', 'Tspan13', 'Ogt', 'Rtn4rl1', 'Arid4a', 'Il6st', 'Pten', 'Nuak1', 'Son', 'Ash1l',
    'Cd37', 'Lrrc3', 'Tra2b', 'Zfp644', 'Il17ra', 'Prex1', 'Zswim6', 'AC149090.1', 'Nfia',
    'Mtus1', 'Ctsf', 'Csnk1e', 'Dock8', 'Man1a', 'Fbrsl1', 'Ets1', 'Srrm2', 'Dusp7',
    'Arhgap17', 'Ppp1r21', 'Arap3', 'Hook3', 'Pik3cg', 'Pnrc1', 'Gna15', 'Zkscan3', 'Mid1',
    'Arhgap45', 'Sft2d2', 'Fam91a1', 'Ppp1r10', 'Aftph', 'Eng', 'Snn', 'Sft2d1', 'Mfap3',
    'Itga6', 'Kmt2c', 'Abi3', 'Rptor', 'Ssbp2', 'Plxdc2', 'Fus', 'Nr3c1', 'Tubgcp5', 'Med13l',
    'Vgll4', 'Clec4a3', 'Bin2', 'Atp6v0a2', 'Map3k8', 'Icam1', 'Fcgr3', 'Tle3', 'Klk8',
    'Rab6b', 'Slc12a6', 'Tspan14', 'Clasp2', 'Fam49b', 'Mkln1', 'BC035044', 'Plcl2', 'Akna',
    'Dleu2', 'Tet3', 'Senp2', 'Trio', 'Ccng2', 'Tsc22d2', 'Rbm26', 'Cyb561a3', 'Rnf167',
    'Dnajb4', 'B430010I23Rik', 'Ppp1r15a', 'Rab3il1', '9930111J21Rik2', 'Heatr5a', 'Nfkbia',
    'Cep68', 'Man1a2', 'Slc40a1', 'Pip4p1', 'Krit1', 'Zfp292', 'Kif21b', 'Myo1b', 'Cyth4',
    'Cd2ap', 'Sgce', 'Klf6', 'Cdk5r1', 'Csad', 'Ttc14', 'Phip', 'Cbl', 'Mef2c', 'Commd8',
    'C5ar2', 'Ep300', 'Gng10', 'Zfp90', 'Wasf2', 'Itpripl1', 'Rhobtb1', 'Gp9', 'Kdm7a',
    'Ankrd11', 'Tln2', 'Sla', 'Cybc1', 'Resf1', 'Foxp1', 'Srsf7', 'Fermt3', 'Rlim', 'Gtf2h2',
    'Tcf4', 'Pid1', 'Rapgef6', 'Fam168a', 'Tmem63a', 'Slc36a1', 'Rsrp1', 'Csf3r', 'Stard9',
    'Inka1', 'Irf2bp2', 'Orai1', 'Lst1', '1810026B05Rik', 'Mafb', 'Gm34589', 'Prpf38b',
    'Map11', 'Taok1', 'Tmcc1', 'Gm47283', 'Zfp869', 'Kdm2b', 'Msl1', 'Cxxc5', 'Zfp871',
    'Bclaf1', 'Pura', 'Chd2', 'Akr1b10', 'Snta1', 'Fli1', 'Mapk14', 'Ccnt2', 'Chst7', 'Lnpep',
    'Pum2', 'Smg1', 'Gdi1', 'Kmt2a', 'Tnrc6c', 'Hck', 'Ldlrad4', 'Gcnt1', 'Tut7', 'Rreb1',
    'Setd2', 'Tmem245', 'Ddx50', 'Lyl1', 'Itch', 'Kansl1', 'Sort1', 'P3h2', 'Lpin2', 'Tasor',
    'Tia1', 'Mgat1', 'Ppp3ca', 'Atf7ip', 'Kat6a', 'Acp2', 'Mlxipl', 'Pcf11', 'Zmiz1',
    'Dennd1c', 'Bmf', 'Syngr2', 'Adamts1', 'Taz', 'Tmem170b', 'Nipbl', 'Neurl1a', 'Itgav',
    'Sh2b3', 'Ckb', 'Tcf7l2', 'Rb1cc1', 'Peli1', 'Fcho2', 'Ccdc152', 'Msl2', 'Gm37494',
    'Sf3b1', 'Tut4', 'Phyhd1', 'Lgmn', 'Gm10790', 'Lrrk1', 'Cebpz', 'Ttc28', 'Gatad2b',
    'Arid3b'
]

reactive_microglia_genes = [
    'Gm42418', 'Lpl', 'Gpnmb', 'Lgals3', 'Lyz2', 'Spp1', 'Apoc1', 'Fabp5', 'Igf1', 'Lars2',
    'Ftl1', 'Apoe', 'Cst7', 'Fth1', 'Cd63', 'Cstb', 'Wfdc17', 'AY036118', 'Lilrb4a', 'Apoc4',
    'Rpl35', 'Ms4a7', 'Ctla2a', 'Pld3', 'Ctsd', 'Ctsb', 'Aldoa', 'Cxcl2', 'Prdx1', 'Mif',
    'Anxa5', 'Gm1673', 'Apoc2', 'Vim', 'Ch25h', 'Cxcl14', 'Rps28', 'Rpl38', 'Lgals1', 'Fn1',
    'Capg', 'Plin2', 'Csf1', 'Syngr1', 'Timp2', 'Ctsz', 'Cd68', 'Cd52', 'mt-Nd4l', 'Fxyd5',
    'Crip1', 'Cd9', 'Rps29', 'Cybb', 'Rpl41', 'mt-Atp8', 'Atp6v0c', 'Fam20c', 'Tyrobp', 'Gapdh',
    'Bcl2a1b', 'Atox1', 'Ccl4', 'Cxcl16', 'Tmsb10', 'Ndufa1', 'Ccl6', 'Cyba', 'Gng5', 'Pkm',
    'Gabarap', 'Gnas', 'Mt1', 'Myl6', 'Bola2', 'Rps2', 'Npc2', 'Uba52', 'Aplp2', 'Ucp2',
    'Ccl3', 'Mbp', 'Serf2', 'Tmem256', 'Creg1', 'Akr1a1', 'Hif1a', 'Rps21', 'Tnf', 'Bcl2a1d',
    'Selenow', 'Rab7b', 'Myo1e', 'Rpl36', 'Sec61g', 'Rpl37', 'Rpl37a', 'Cd72', 'Rps26', 'Pgam1',
    'Ctsa', 'Itgax', 'Iqgap1', 'Ftl1-ps1', 'Bri3', 'Rpl36a', 'Cotl1', 'Serpinb6a', 'Rps27',
    'Rpl39', 'Lipa', 'Cox7c', 'Fabp3', 'mt-Nd5', 'Pgk1', 'Rpl23', 'Atp6v1f', 'Tmsb4x',
    'Sh3bgrl3', 'Nme2', 'Rpl10', 'Rap2b', 'Atp6v1a', 'Tlr2', 'Tpi1', 'Ctsl', 'Plek', 'Ctla2b',
    'Cox8a', 'Ank', 'Lrpap1', 'Fcer1g', 'Rhoc', 'Rps19', 'Cox6a2', 'Vat1', 'Rpl22l1', 'Tomm6',
    'Sem1', 'Atp5k', 'Cox6c', 'Cox6b1', 'Ndufb1-ps', 'Atp6v1g1', 'Aprt', 'Atp5md', 'Gng12',
    'Axl', 'Anxa2', 'Atp5e', 'Rpl32', 'Rpl24', 'Nceh1', 'Rpl10a', 'Rpl6', 'Aldh2', 'Ssr4',
    'Rps18', 'Tomm7', 'Rps20', 'Msrb1', 'Npm1', 'Slc15a3', 'Pianp', 'Glmp', 'Tpd52', 'Rpl22',
    'Rps23', 'Crlf2', 'Lilr4b', 'Gusb', 'Uqcr11', 'Srgn', 'Gyg', 'Cox4i1', 'Rpl31', 'Tpt1',
    'Rpl19', 'Atp6v0e', 'Clta', 'Grn', 'Lamp1', 'Atp5mpl', 'Chchd2', 'Ndufa2', '1810058I24Rik',
    'Rpsa', 'Ldha', 'Rplp2', 'Uqcrq', 'Colec12', 'Rplp0', 'Txndc17', 'Romo1', 'Id2', 'Nsa2',
    'Rps12', 'Ms4a6c', 'Atp5l', 'Dpp7', 'Gpx4', 'Emp3', 'Rpl3', 'mt-Co1', 'Psmb6', 'Cd48',
    'H2-D1', 'Gm11808', 'S100a1', 'Rps25', 'Eef1a1', 'Rpl28', 'Eno1', 'Cops9', 'Rpl12', 'Rpl34',
    'Tbca', 'Ubl5', 'Hspe1', 'Rpl27a', 'Naca', 'Plaur', 'Gla', 'Cox7b', 'Ndufv3', 'Sec61b',
    'Bax', 'Hexa', 'Cfl1', 'H2afj', 'Rps24', 'Trem2', 'Soat1', 'Gas6', 'Elob', 'Rps8', 'Rplp1',
    'Cpd', 'Rps16', 'Gm10076', 'C3ar1', 'Rpl17', 'Pebp1', 'Arpc1b', 'Fundc2', 'Flt1', 'Sdcbp',
    'Tmed3', 'Ndufa13', 'Rpl7a', 'Atp5j2', 'Prdx5', 'Myo5a', 'Calm1', 'Rps5', 'Arpc3', 'Echs1',
    'Il4i1', 'Calr', 'Rpl11', 'Ndufa7', 'Sqstm1', 'Rpl26', 'Hcst', 'Nme1', 'Atp5g2', 'Serpine1',
    'Rps15a', 'Uqcrh', 'Cox5b', 'Fau', 'Abca1', 'Uap1l1', 'Rpl35a', 'Rpl15', 'Gpi1', 'Ost4',
    'Rps10', 'Tmem160', 'Csf2ra', 'Rps6', 'Rps13', 'Gpr65', '0610012G03Rik', 'Rpl14', 'Atf3',
    'Adssl1', 'Dab2', 'Eef1g', 'Ndufa3', 'Chst2', 'Glipr1', 'Rps27l', 'mt-Nd3', 'Dtnbp1',
    'Cox17', 'Sh3bgrl', 'Vps13c', 'Cox6a1', 'Phlda1', 'Rps17', 'Gpr137b', 'Alcam', 'Rpl30',
    'Rgs1', 'Sat1', 'Hcar2', 'Mpeg1', 'Ndufb2', 'Ybx1', 'Gsto1', 'Rack1', 'Atp5j', 'Rps7',
    'Rps27a', 'Tceal9', 'Uqcr10', 'Amdhd2', 'Arl11', 'Bcl2a1a', 'Ndufs5', 'C1qa', 'Fuca1',
    'Grcc10', 'Blvrb', 'Rpl36al', 'Mt2', 'Litaf', 'Atp5h', 'Eef1b2', 'Rpl9', 'Neat1', 'Ndufc1',
    'Hsp90ab1', 'Arpc5', 'Pdgfa', 'Pet100', 'Cox7a2', 'Ndufb7', 'B2m', 'Rpl13', 'Map1lc3b',
    'Ptms', 'Lamtor4', 'Ninj1', 'Fblim1', 'Sdf2l1', 'Tmem258', 'Arpc2', 'Tma7', 'Atp5g1',
    'Abcg1', 'Rpl27', 'Slc25a3', 'Ctnnb1', 'Gde1', 'Cspg4', 'Apbb2', 'Lamtor2', 'mt-Co2',
    'Smpdl3a', 'Rnasek', 'Sars', 'Vps29', 'Akirin1', 'Uqcrb', 'Edf1', 'Hist1h2bc', 'Atp6v1e1',
    'Tnfaip2', 'Cdkn1a', 'Rpl29', 'Dusp3', 'Pfdn5', 'Ggh', 'Scd2', 'Rpl8', 'Timm8b', 'Mdfic',
    'Vdac2', 'Adam8', 'Micos10', 'Ubxn1', 'Prdx4', 'Rps3', 'Eif3i', 'Samsn1', 'Eif2s2',
    'Mrpl52', 'Rpl21', 'Eif3k', 'Anxa4', 'Cd83', 'Il1b', 'Ndufb8', 'Actg1', 'Bhlhe40',
    'Rpl18', 'Btf3', 'Pkib', 'Rps14', 'Atp6v1c1', 'Tmem205', 'Ppt1', 'Dhrs3', 'Ndufb9',
    'Gas2l3', 'Fcrls', 'Naglu', 'Capns1', 'Scpep1', 'Folr2', 'Fis1', 'Gngt2', 'Tent5c',
    'Msr1', 'Plau', 'Pf4', 'Smdt1', 'Atpif1', 'Ly9', 'Pla2g7', 'Hint1', 'Mrpl33', 'Tomm20',
    'Rpl7', 'Bsg', 'Ppib', 'Nfe2l2', 'Cd14', 'Uqcc2', 'Rtcb', 'Rftn1', 'Eif1', 'Aif1',
    'Mrpl54', 'Dbi', 'Hebp1', 'mt-Co3', 'Dstn', '1700003F12Rik', 'Sash1', 'Fkbp2', 'Bag1',
    'Ube2l3', 'Calm3', 'St13', 'Gas5', 'Polr1d', 'Mrps21', 'Rps9', 'Mien1', 'Fnip2', 'Nedd8',
    'Ass1', 'Txn1', 'Eif4a1', 'Rps11', 'Rap2a', 'Got1', 'Micos13', 'Prelid1', 'Cyp4f18',
    'Clic4', 'Snhg6', 'Trappc2l', 'Llph', 'AB124611', 'Srp14', 'Snx3', 'Pgls', 'Ppia',
    'Cox5a', 'Eif5a', 'Mtln', 'Pltp', 'Ndufa4', 'Osm', 'S100a13', 'Renbp', 'Gba', 'Itgb2',
    'Esd', 'Tpp1', 'Il3ra', 'Atp5c1', 'Rai14', 'Cd22', 'Cndp2', 'Swi5', 'Ndufa6', 'Chmp2a',
    'Plekhm2', 'Ahnak2'
]

gene_expr_frac = (adata.X > 0).sum(axis=0) / adata.n_obs
expressed_genes = adata.var_names[gene_expr_frac > 0.05]

# Filter gene lists
homeostatic_filtered = [gene for gene in homeostatic_microglia_genes if gene in expressed_genes]
reactive_filtered = [gene for gene in reactive_microglia_genes if gene in expressed_genes]

sc.tl.score_genes(adata, gene_list=homeostatic_filtered, score_name='homeostatic_microglia_score')
sc.tl.score_genes(adata, gene_list=reactive_filtered, score_name='reactive_microglia_score')

# Microglia reactivity index
adata.obs['microglia_reactivity_index'] = (
    adata.obs['reactive_microglia_score'] - adata.obs['homeostatic_microglia_score']
)

# Viz
sc.pl.violin(
    adata,
    keys=['homeostatic_microglia_score', 'reactive_microglia_score', 'microglia_reactivity_index'],
    groupby='group',
    jitter=True,
    stripplot=True
)

sc.pl.matrixplot(adata, var_names = ['homeostatic_microglia_score', 'reactive_microglia_score'], groupby = 'group', standard_scale = 'var',
                save = 'microglia-matrixplot.pdf')

# Ensure clean data
plot_data = adata.obs.loc[
    adata.obs['group'].isin(['Sham-GFP', 'Sham-VEGFC', 'TBI-GFP', 'TBI-VEGFC']),
    ['group', 'microglia_reactivity_index']
].dropna()

# Plot
plt.figure(figsize=(6, 5))
ax = sns.violinplot(
    data=plot_data,
    x='group',
    y='microglia_reactivity_index',
    palette='Set2',
    inner='box'
)

# Add vertical space without shifting violins
y_min, y_max = ax.get_ylim()
ax.set_ylim(y_min, y_max + 0.5)  # just adds buffer above violins

# Titles and labels
ax.set_title('')
ax.set_ylabel('Microglia Reactivity Index')
ax.set_xlabel('')

# Layout
plt.tight_layout()
plt.grid(False)
pdf_path = os.path.join(fig_dir, "microglia_reactivity_violin.pdf")
plt.savefig(pdf_path, format='pdf')
plt.show()

from itertools import product
from scipy.stats import mannwhitneyu

def cliffs_delta(x, y):
    nx, ny = len(x), len(y)
    more = sum(1 for a, b in product(x, y) if a > b)
    less = sum(1 for a, b in product(x, y) if a < b)
    return (more - less) / (nx * ny)

# Ensure group labels are strings
adata.obs['group'] = adata.obs['group'].astype(str)
score_key = 'microglia_reactivity_index'

### 1. TBI-GFP vs Sham-GFP (expecting TBI-GFP > Sham-GFP)
g1_vals = adata.obs.loc[adata.obs['group'] == 'TBI-GFP', score_key].dropna().values
g2_vals = adata.obs.loc[adata.obs['group'] == 'Sham-GFP', score_key].dropna().values

stat1, p1 = mannwhitneyu(g1_vals, g2_vals, alternative='greater')
delta1 = cliffs_delta(g1_vals, g2_vals)

print("TBI-GFP vs Sham-GFP")
print(f"One-sided Mann–Whitney U p = {p1:.4f}")
print(f"Cliff's Delta = {delta1:.3f}\n")

### 2. TBI-GFP vs TBI-VEGFC (expecting TBI-GFP > TBI-VEGFC — rescue)
g1_vals = adata.obs.loc[adata.obs['group'] == 'TBI-GFP', score_key].dropna().values
g2_vals = adata.obs.loc[adata.obs['group'] == 'TBI-VEGFC', score_key].dropna().values

stat2, p2 = mannwhitneyu(g1_vals, g2_vals, alternative='greater')
delta2 = cliffs_delta(g1_vals, g2_vals)

print("TBI-GFP vs TBI-VEGFC")
print(f"One-sided Mann–Whitney U p = {p2:.4f}")
print(f"Cliff's Delta = {delta2:.3f}\n")

### 3. Sham-GFP vs TBI-VEGFC (expecting TBI-VEGFC > Sham-GFP)
g1_vals = adata.obs.loc[adata.obs['group'] == 'TBI-VEGFC', score_key].dropna().values
g2_vals = adata.obs.loc[adata.obs['group'] == 'Sham-VEGFC', score_key].dropna().values

stat3, p3 = mannwhitneyu(g1_vals, g2_vals, alternative='greater')
delta3 = cliffs_delta(g1_vals, g2_vals)

print("TBI-VEGFC vs Sham-VEGFC")
print(f"One-sided Mann–Whitney U p = {p3:.4f}")
print(f"Cliff's Delta = {delta3:.3f}")
